Hide play button in media player if there is not media to control
If the media player does not have media to control, the
play/pause/stop, previous, and next buttons will be hidden. In order
for the playback buttons to be shown, the status of the media player
must be playing, paused, or unknown. This way, if the status is On, the
playback buttons will be hidden.