make more-info- tags in dynamicContentUpdater lowercase (#211)
When trying to understand the source tree using `git grep` the fact
that the critical function which actually creates the more-info-
panels used MORE-INFO, makes it actually hard to find. Rework the
logic so that it's all managed in lower case for greater
discoverability of hunting through code.

Because the stateMoreInfoType comes from home-assistant domains, this
should always be lower case, so no explicit lower casing is needed.